button: Support Block API Version 2
