+++ blocks/src/block/button/block.json
@@ -1,0 +1,56 @@
+{
+  "apiVersion": 2,
+  "name": "cocoon-blocks/button-1",
+  "category": "cocoon-block",
+  "keywords": [ "button", "btn" ],
+
+  "attributes": {
+    "content": {
+      "type": "string",
+      "default": "Button"
+    },
+    "url": {
+      "type": "string",
+      "default": ""
+    },
+    "target": {
+      "type": "string",
+      "default": "_self"
+    },
+    "size": {
+      "type": "string",
+      "default": ""
+    },
+    "isCircle": {
+      "type": "boolean",
+      "default": false
+    },
+    "isShine": {
+      "type": "boolean",
+      "default": false
+    },
+    "align": {
+      "type": "string"
+    },
+    "backgroundColor": {
+      "type": "string"
+    },
+    "textColor": {
+      "type": "string"
+    },
+    "borderColor": {
+      "type": "string"
+    },
+    "customBorderColor": {
+      "type": "string"
+    },
+    "fontSize": {
+      "type": "string"
+    }
+  },
+  "supports": {
+    "align": [ "left", "center", "right" ],
+    "className": false,
+    "customClassName": true
+  }
+}
