+++ blocks/src/init.php
@@ -278,6 +278,7 @@ add_action('init', function () {
 require_once abspath(__FILE__).'block/balloon/index.php';
 require_once abspath(__FILE__).'block/blank-box/index.php';
 require_once abspath(__FILE__).'block/blogcard/index.php';
+require_once abspath(__FILE__).'block/button/index.php';
 require_once abspath(__FILE__).'block/icon-box/index.php';
 require_once abspath(__FILE__).'block/icon-list/index.php';
 require_once abspath(__FILE__).'block/info-box/index.php';
